requirements_base.txt
otree==2.5.3
django-recaptcha==2.0.5
